+++ R/mod_safetyGraphicsUI.R
@@ -2,7 +2,7 @@
 #'
 #'
 #' @param id module ID
-#' @param meta data frame containing the metadata for use in the app. See the preloaded file (\code{?safetyGraphics::meta}) for more data specifications and details. Defaults to \code{safetyGraphics::meta}. 
+#' @param meta data frame containing the metadata for use in the app. 
 #' @param domainData named list of data.frames to be loaded in to the app.
 #' @param mapping data.frame specifying the initial values for each data mapping. If no mapping is provided, the app will attempt to generate one via \code{detectStandard()}
 #' @param standards a list of information regarding data standards. Each list item should use the format returned by safetyGraphics::detectStandard.
@@ -12,16 +12,12 @@
 #' @export
 
 safetyGraphicsUI <- function(id, meta, domainData, mapping, standards){
-    #read css from pacakge
     ns<-NS(id)
-    app_css <- NULL
-    for(lib in .libPaths()){
-        if(is.null(app_css)){
-            css_path <- paste(lib,'safetyGraphics', 'www','index.css', sep="/")
-            if(file.exists(css_path)) app_css <-  HTML(readLines(css_path))
-        }
-    }
-    
+
+    #read css from package
+    css_path <- system.file("www","index.css", package = "safetyGraphics")
+    app_css <-  HTML(readLines(css_path))
+
     #script to append population badge nav bar
     participant_badge<-tags$script(
         HTML(