+++ inst/deploy/app.R
@@ -1,0 +1,5 @@
+# Launch the ShinyApp (Do not remove this comment)
+# To deploy, run: rsconnect::deployApp()
+# Or use the blue button on top of this file
+library(safetyGraphics) # or use devtools::install_github() to load dev version
+safetyGraphics::safetyGraphicsApp(runNow = FALSE) # add parameters here (if any)