+++ tests/testthat/module_examples/filterTab/app.R
@@ -5,6 +5,14 @@ library(safetyGraphics)
 
 #reactlogReset()
 
+meta <- rbind(
+  safetyCharts::meta_labs,
+  safetyCharts::meta_aes,
+  safetyCharts::meta_dm,
+  safetyCharts::meta_hepExplorer
+)
+
+
 ui <- tagList(
     shinyjs::useShinyjs(),
     tags$head(